+++ ExpandWorld/Spawn.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 
 namespace ExpandWorld;
@@ -84,5 +85,48 @@ public class Spawn
     // This is used for the top level objects so dungeon generation is done by vanilal code.
     Data.CleanGhostInit(go);
     return go;
+  }
+
+  private static List<Tuple<float, string>> ParseSwap(IEnumerable<string> items)
+  {
+    var total = 0f;
+    return items.Select(s => s.Split(':')).Select(s =>
+    {
+      var weight = Parse.Float(s, 1, 1f);
+      total += weight;
+      return Tuple.Create(weight, s[0]);
+    }).ToList().Select(t => Tuple.Create(t.Item1 / total, t.Item2)).ToList();
+  }
+
+  public static Dictionary<string, List<Tuple<float, string>>> LoadSwaps(string[] objectSwap)
+  {
+    var swaps = objectSwap.Select(Data.ToList)
+        .ToDictionary(arr => arr[0], arr => Spawn.ParseSwap(arr.Skip(1)));
+    foreach (var kvp in swaps)
+    {
+      foreach (var swap in kvp.Value)
+        BlueprintManager.Load(swap.Item2, "");
+    }
+    return swaps;
+  }
+
+  public static Dictionary<string, ZDO?> LoadData(string[] objectData)
+  {
+    return objectData.Select(Data.ToList).ToDictionary(arr => arr[0], arr => Data.ToZDO(arr[1]));
+  }
+
+  public static string RandomizeSwap(List<Tuple<float, string>> swaps)
+  {
+    if (swaps.Count == 0)
+      return "";
+    if (swaps.Count == 1)
+      return swaps[0].Item2;
+    var rng = UnityEngine.Random.value;
+    foreach (var swap in swaps)
+    {
+      rng -= swap.Item1;
+      if (rng <= 0f) return swap.Item2;
+    }
+    return swaps[swaps.Count - 1].Item2;
   }
 }