+++ ExpandWorld/dungeon/DungeonSpawning.cs
@@ -4,28 +4,47 @@ using System.Reflection.Emit;
 using HarmonyLib;
 using UnityEngine;
 
-namespace ExpandWorld;
+namespace ExpandWorld.Dungeon;
 
-///<summary>Dungeon level customization</summary>
 [HarmonyPatch(typeof(DungeonGenerator))]
-public class DungeonSpawning
+public class Spawner
 {
   ///<summary>Dungeon doesn't know its location so it must be tracked manually.</summary>
   public static ZoneSystem.ZoneLocation? Location = null;
 
-  ///<summary>Implements object data and swapping from location data.</summary>
-  static GameObject CustomObject(GameObject prefab, Vector3 position, Quaternion rotation)
+  public static Dictionary<string, FakeDungeonGenerator> Generators = new();
+  public static string RoomName = "";
+  public static string DungeonName = "";
+
+  public static bool TryGetSwap(string dungeon, string prefab, out string swapped)
   {
-    // Some mods cause client side dungeon reloading.
-    // In this case location information is not available.
-    // Revert to the default behaviour as fail safe (no swaps or blueprint won't be available).
-    if (Location == null) return UnityEngine.Object.Instantiate<GameObject>(prefab, position, rotation);
+    swapped = "";
+    if (!Generators.TryGetValue(dungeon, out var gen)) return false;
+    if (!gen.m_objectSwaps.TryGetValue(prefab, out var swaps)) return false;
+    swapped = Spawn.RandomizeSwap(swaps);
+    return true;
+  }
+
+  public static ZDO? DataOverride(string dungeon, string prefab)
+  {
+    if (!Generators.TryGetValue(dungeon, out var gen)) return null;
+    if (!gen.m_objectData.TryGetValue(prefab, out var data)) return null;
+    return data;
+  }
+
+  ///<summary>Implements object data and swapping from location data.</summary>
+  static GameObject CustomObject(DungeonGenerator dg, GameObject prefab, Vector3 position, Quaternion rotation)
+  {
+    // Some mods cause client side dungeon reloading. In this case, no data is available.
+    // Revert to the default behaviour as a fail safe.
+    if (Helper.IsClient()) return UnityEngine.Object.Instantiate<GameObject>(prefab, position, rotation);
     BlueprintObject bpo = new(Utils.GetPrefabName(prefab), position, rotation, prefab.transform.localScale, "", null, 1f);
-    var locName = Location?.m_prefabName ?? "";
-    if (LocationSpawning.TryGetSwap(locName, bpo.Prefab, out var objName))
+    if (TryGetSwap(dg.name, bpo.Prefab, out var objName))
+      bpo.Prefab = objName;
+    if (RoomSpawning.TryGetSwap(RoomName, bpo.Prefab, out objName))
       bpo.Prefab = objName;
 
-    var obj = Spawn.BPO(locName, bpo, LocationSpawning.DataOverride, null);
+    var obj = Spawn.BPO(dg.name, bpo, DataOverride, null);
     return obj ?? LocationSpawning.DummySpawn;
   }
 
@@ -34,6 +53,7 @@ public class DungeonSpawning
     var instantiator = typeof(UnityEngine.Object).GetMethods().First(m => m.Name == nameof(UnityEngine.Object.Instantiate) && m.IsGenericMethodDefinition && m.GetParameters().Skip(1).Select(p => p.ParameterType).SequenceEqual(new[] { typeof(Vector3), typeof(Quaternion) })).MakeGenericMethod(typeof(GameObject));
     return new CodeMatcher(instructions)
       .MatchForward(false, new CodeMatch(OpCodes.Call, instantiator))
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Ldarg_0))
       .Set(OpCodes.Call, Transpilers.EmitDelegate(CustomObject).operand)
       .InstructionEnumeration();
   }
@@ -46,4 +66,51 @@ public class DungeonSpawning
   [HarmonyPatch(nameof(DungeonGenerator.PlaceDoors)), HarmonyTranspiler]
   static IEnumerable<CodeInstruction> DungeonDoors(IEnumerable<CodeInstruction> instructions) => Transpile(instructions);
 
+
+  [HarmonyPatch(nameof(DungeonGenerator.Generate), typeof(ZoneSystem.SpawnMode)), HarmonyPrefix]
+  static void Generate(DungeonGenerator __instance)
+  {
+    if (Location == null) return;
+    var dungeonName = Utils.GetPrefabName(__instance.gameObject);
+    var locName = Location?.m_prefabName ?? "";
+    if (LocationLoading.LocationData.TryGetValue(locName, out var data) && data.dungeon != "")
+      dungeonName = data.dungeon;
+    Override(__instance, dungeonName);
+  }
+
+
+  public static void Override(DungeonGenerator dg, string name)
+  {
+    if (!Generators.TryGetValue(name, out var data)) return;
+    dg.name = name;
+    dg.m_algorithm = data.m_algorithm;
+    dg.m_zoneSize = data.m_zoneSize;
+    dg.m_alternativeFunctionality = data.m_alternativeFunctionality;
+    dg.m_campRadiusMax = data.m_campRadiusMax;
+    dg.m_campRadiusMin = data.m_campRadiusMin;
+    dg.m_doorChance = data.m_doorChance;
+    dg.m_doorTypes = data.m_doorTypes;
+    dg.m_maxRooms = data.m_maxRooms;
+    dg.m_minRooms = data.m_minRooms;
+    dg.m_maxTilt = data.m_maxTilt;
+    dg.m_minAltitude = data.m_minAltitude;
+    dg.m_minRequiredRooms = data.m_minRequiredRooms;
+    dg.m_requiredRooms = data.m_requiredRooms;
+    dg.m_themes = data.m_themes;
+    dg.m_gridSize = data.m_gridSize;
+    dg.m_tileWidth = data.m_tileWidth;
+    dg.m_spawnChance = data.m_spawnChance;
+    dg.m_perimeterSections = data.m_perimeterSections;
+    dg.m_perimeterBuffer = data.m_perimeterBuffer;
+    dg.m_useCustomInteriorTransform = data.m_useCustomInteriorTransform;
+  }
+  [HarmonyPatch(typeof(DungeonGenerator), nameof(DungeonGenerator.SetupAvailableRooms)), HarmonyPostfix]
+  public static void SetupAvailableRooms(DungeonGenerator __instance)
+  {
+    var name = Utils.GetPrefabName(__instance.gameObject);
+    if (!Generators.TryGetValue(name, out var gen)) return;
+    if (gen.m_excludedRooms.Count == 0) return;
+    DungeonGenerator.m_availableRooms = DungeonGenerator.m_availableRooms.Where(room => !gen.m_excludedRooms.Contains(room.m_room.name)).ToList();
+
+  }
 }