+++ publish/CHANGELOG.md
@@ -1,3 +1,9 @@
+- v1.41
+  - Adds new fields `bounds` and `objectSwap` to the `expand_dungeons.yaml` file.
+  - Adds new fields `objectData` and `objectSwap` to the `expand_dungeons.yaml` file.
+  - Changes the fields `objectSwap` and `objectData` in the `expand_locations.yaml` to not affect dungeons.
+  - Fixes custom dungeons not working.
+  
 - v1.40
   - Adds blueprint dungeons rooms.
   - Fixes dungeons not working when first starting a single player world and then joining a multiplayer world.